requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=0_gnu
backcall=0.1.0=py_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
cycler=0.10.0=py_2
dbus=1.13.6=he372182_0
decorator=4.4.2=py_0
expat=2.2.9=he1b5a44_2
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.1=he06d7ca_0
gettext=0.19.8.1=hc5be6a0_1002
glib=2.58.3=py37he00f558_1004
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
icu=64.2=he1b5a44_1
imagecodecs-lite=2019.12.3=py37h8f50634_0
ipython_genutils=0.2.0=py37_0
jpeg=9c=h14c3975_1001
kiwisolver=1.2.0=py37h99015e2_0
ld_impl_linux-64=2.34=h53a641e_0
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libclang=9.0.1=default_hde54327_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.3.0=hdf63c60_5
libgomp=9.2.0=h24d8f2e_2
libiconv=1.15=h516909a_1006
liblapack=3.8.0=14_openblas
libllvm9=9.0.1=hc9558a2_0
libopenblas=0.3.7=h5ec1e0e_6
libpng=1.6.37=hed695b0_1
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc7e4089_6
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lz4-c=1.8.3=he1b5a44_1001
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py37h30547a4_0
ncurses=6.1=hf484d3e_1002
nspr=4.25=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.18.1=py37h8960a57_1
olefile=0.46=py37_0
openssl=1.1.1=h7b6447c_0
pandas=0.25.3=py37he6710b0_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=py_2
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py37_0
pygments=2.6.1=py_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37hcca6a23_1
python=3.7.6=h8356626_5_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
qt=5.12.5=hd8c4c69_1
readline=8.0=hf8c457e_0
setuptools=46.1.3=py37hc8dfbb8_0
simpleitk=1.2.4=py37hf484d3e_0
six=1.14.0=py_1
sqlite=3.30.1=hcee41ef_0
tifffile=2020.2.16=py_0
tk=8.6.10=hed695b0_0
tornado=6.0.4=py37h8f50634_1
traitlets=4.3.3=py37_0
wcwidth=0.1.9=py_0
wheel=0.34.2=py_1
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.5=h516909a_0
zlib=1.2.11=h516909a_1006
zstd=1.4.4=h3b9ef0a_2
